[MODIFIED] Se modularizaron los endpoints
