+++ backend/api/controller/UsuarioController.py
@@ -21,30 +21,4 @@ def login():
     data = request.get_json()
     usuario = data.get('usuario')
     password = data.get('password')
-
-    if not usuario or not password:
-        return jsonify({"error": "Faltan credenciales", "codigo": 400}), 400
-
-    conexion = obtener_conexion()
-    cursor = conexion.cursor()
-
-    cursor.execute("""
-        SELECT id_usuario, usuario, email, tipo_usuario, activo
-        FROM usuarios
-        WHERE usuario = :1 AND password = :2
-    """, (usuario, password))
-
-    fila = cursor.fetchone()
-    cursor.close()
-    conexion.close()
-
-    if fila:
-        return jsonify({
-            "id_usuario": fila[0],
-            "usuario": fila[1],
-            "email": fila[2],
-            "tipo_usuario": fila[3],
-            "activo": fila[4]
-        }), 200
-    else:
-        return jsonify({"error": "Usuario o contraseña incorrectos", "codigo": 401}), 401
+    return UsuarioService.login(usuario, password)
